+++ axle-core/src/main/scala/axle/algebra/distance/Manhattan.scala
@@ -1,4 +1,4 @@
-package axle.ml.distance
+package axle.algebra.distance
 
 import axle.algebra.LinearAlgebra
 import axle.syntax.linearalgebra.matrixOps
